Split the app/new behavior between config and enterprise
